so many changes, i was going to split it up into many commits but i realized there would be at least 12 discrete commits

+++ source/emu/hw/spi/device/package.d
@@ -9,7 +9,7 @@ public {
 import util;
 
 abstract class SPIDevice {
-    Half write(Byte b);
+    Byte write(Byte b);
     void chipselect_fall();
     void chipselect_rise();
 }
